+++ src/equation/cg.cpp
@@ -77,10 +77,20 @@ int equation::CG<MATRIX, T>::monolish_CG(MATRIX &A, vector<T> &x,
   logger.solver_out();
   return MONOLISH_SOLVER_MAXITER;
 }
+template int equation::CG<matrix::Dense<double>, double>::monolish_CG(
+    matrix::Dense<double> &A, vector<double> &x, vector<double> &b);
+template int equation::CG<matrix::Dense<float>, float>::monolish_CG(
+    matrix::Dense<float> &A, vector<float> &x, vector<float> &b);
+
 template int equation::CG<matrix::CRS<double>, double>::monolish_CG(
     matrix::CRS<double> &A, vector<double> &x, vector<double> &b);
 template int equation::CG<matrix::CRS<float>, float>::monolish_CG(
     matrix::CRS<float> &A, vector<float> &x, vector<float> &b);
+
+template int equation::CG<matrix::LinearOperator<double>, double>::monolish_CG(
+    matrix::LinearOperator<double> &A, vector<double> &x, vector<double> &b);
+template int equation::CG<matrix::LinearOperator<float>, float>::monolish_CG(
+    matrix::LinearOperator<float> &A, vector<float> &x, vector<float> &b);
 
 ///
 
@@ -97,8 +107,18 @@ int equation::CG<MATRIX, T>::solve(MATRIX &A, vector<T> &x, vector<T> &b) {
   logger.solver_out();
   return ret; // err code
 }
+template int equation::CG<matrix::Dense<double>, double>::solve(
+    matrix::Dense<double> &A, vector<double> &x, vector<double> &b);
+template int equation::CG<matrix::Dense<float>, float>::solve(
+    matrix::Dense<float> &A, vector<float> &x, vector<float> &b);
+
 template int equation::CG<matrix::CRS<double>, double>::solve(
     matrix::CRS<double> &A, vector<double> &x, vector<double> &b);
 template int equation::CG<matrix::CRS<float>, float>::solve(
     matrix::CRS<float> &A, vector<float> &x, vector<float> &b);
+
+template int equation::CG<matrix::LinearOperator<double>, double>::solve(
+    matrix::LinearOperator<double> &A, vector<double> &x, vector<double> &b);
+template int equation::CG<matrix::LinearOperator<float>, float>::solve(
+    matrix::LinearOperator<float> &A, vector<float> &x, vector<float> &b);
 } // namespace monolish