+++ src/equation/none.cpp
@@ -17,6 +17,11 @@ void equation::none<MATRIX, T>::create_precond(MATRIX &A) {
   logger.solver_in(monolish_func);
   logger.solver_out();
 }
+template void equation::none<matrix::Dense<float>, float>::create_precond(
+    matrix::Dense<float> &A);
+template void equation::none<matrix::Dense<double>, double>::create_precond(
+    matrix::Dense<double> &A);
+
 template void equation::none<matrix::CRS<float>, float>::create_precond(
     matrix::CRS<float> &A);
 template void equation::none<matrix::CRS<double>, double>::create_precond(
@@ -39,10 +44,22 @@ void equation::none<MATRIX, T>::apply_precond(const vector<T> &r,
   blas::copy(r, z);
   logger.solver_out();
 }
+template void equation::none<matrix::Dense<float>, float>::apply_precond(
+    const vector<float> &r, vector<float> &z);
+template void equation::none<matrix::Dense<double>, double>::apply_precond(
+    const vector<double> &r, vector<double> &z);
+
 template void
 equation::none<matrix::CRS<float>, float>::apply_precond(const vector<float> &r,
                                                          vector<float> &z);
 template void equation::none<matrix::CRS<double>, double>::apply_precond(
+    const vector<double> &r, vector<double> &z);
+
+template void
+equation::none<matrix::LinearOperator<float>, float>::apply_precond(
+    const vector<float> &r, vector<float> &z);
+template void
+equation::none<matrix::LinearOperator<double>, double>::apply_precond(
     const vector<double> &r, vector<double> &z);
 
 /////
@@ -58,51 +75,18 @@ int equation::none<MATRIX, T>::solve(MATRIX &A, vector<T> &x, vector<T> &b) {
   logger.solver_out();
   return MONOLISH_SOLVER_SUCCESS;
 }
+template int equation::none<matrix::Dense<float>, float>::solve(
+    matrix::Dense<float> &A, vector<float> &x, vector<float> &b);
+template int equation::none<matrix::Dense<double>, double>::solve(
+    matrix::Dense<double> &A, vector<double> &x, vector<double> &b);
+
 template int equation::none<matrix::CRS<float>, float>::solve(
     matrix::CRS<float> &A, vector<float> &x, vector<float> &b);
 template int equation::none<matrix::CRS<double>, double>::solve(
     matrix::CRS<double> &A, vector<double> &x, vector<double> &b);
 
-//////////////////////////////////////////////////////
-// solver set precond /////////////////////////////////
-//////////////////////////////////////////////////////
-template <typename MATRIX, typename T>
-template <class PRECOND>
-void solver::solver<MATRIX, T>::set_create_precond(PRECOND &p) {
-  Logger &logger = Logger::get_instance();
-  logger.util_in(monolish_func);
-  precond.create_precond =
-      std::bind(&PRECOND::create_precond, &p, std::placeholders::_1);
-  logger.util_out();
-}
-
-template void solver::solver<matrix::CRS<double>, double>::set_create_precond(
-    equation::none<matrix::CRS<double>, double> &p);
-template void solver::solver<matrix::CRS<float>, float>::set_create_precond(
-    equation::none<matrix::CRS<float>, float> &p);
-template void solver::solver<matrix::CRS<double>, double>::set_create_precond(
-    equation::Jacobi<matrix::CRS<double>, double> &p);
-template void solver::solver<matrix::CRS<float>, float>::set_create_precond(
-    equation::Jacobi<matrix::CRS<float>, float> &p);
-
-/////
-template <typename MATRIX, typename T>
-template <class PRECOND>
-void solver::solver<MATRIX, T>::set_apply_precond(PRECOND &p) {
-  Logger &logger = Logger::get_instance();
-  logger.util_in(monolish_func);
-  precond.apply_precond =
-      std::bind(&PRECOND::apply_precond, &p, std::placeholders::_1,
-                std::placeholders::_2);
-  logger.util_out();
-}
-
-template void solver::solver<matrix::CRS<double>, double>::set_apply_precond(
-    equation::none<matrix::CRS<double>, double> &p);
-template void solver::solver<matrix::CRS<float>, float>::set_apply_precond(
-    equation::none<matrix::CRS<float>, float> &p);
-template void solver::solver<matrix::CRS<double>, double>::set_apply_precond(
-    equation::Jacobi<matrix::CRS<double>, double> &p);
-template void solver::solver<matrix::CRS<float>, float>::set_apply_precond(
-    equation::Jacobi<matrix::CRS<float>, float> &p);
+template int equation::none<matrix::LinearOperator<float>, float>::solve(
+    matrix::LinearOperator<float> &A, vector<float> &x, vector<float> &b);
+template int equation::none<matrix::LinearOperator<double>, double>::solve(
+    matrix::LinearOperator<double> &A, vector<double> &x, vector<double> &b);
 } // namespace monolish